generator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
beanie==1.23.6
pymongo==4.6.0
google-generativeai==0.8.5
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
requests==2.31.0
python-multipart==0.0.6
openai==1.8.0 